% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function addTrans.R
\name{addTrans}
\alias{addTrans}
\title{This function is used to optimally display the networks with qgraph.
Works with either color and trans a vector of equal length,
or one of the two of length 1.}
\usage{
addTrans(color, trans)
}
\arguments{
\item{color}{color specification for the arrows in the plot}

\item{trans}{transparency: an integer between 0 and 255, 
0 being fully transparent and 255 being fully visable}
}
\value{
a vector with specificatiosn used in qgraph
}
\description{
This function is used to optimally display the networks with qgraph.
Works with either color and trans a vector of equal length,
or one of the two of length 1.
}
\examples{
specs <- addTrans(color = "green", trans=255)
}
