% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerESM.R
\name{centerESM}
\alias{centerESM}
\title{Center variables in ESM data}
\usage{
centerESM(data, subjnr = NULL, addmeans = TRUE, varnames = NULL, center = NULL)
}
\arguments{
\item{data}{data frame}

\item{subjnr}{variable, indicating subjects}

\item{addmeans}{logical indicating whether a vector of person means should be added to the data 
(works only if person centering is requested for that variable).}

\item{varnames}{vector with the names of the variables which are centered.}

\item{center}{vector of same length as varnames that indicates for each variable which centering must be,
applied, either "person" or "grand_mean" or a number. If "person" or "grand_mean" is specified once (default), 
all variables are person or grand-mean centered, respectively. 
If center is NULL no variables are centered and a warning is issued.}
}
\value{
data frame with centered variables, and if requested, person means with suffix "_means".
}
\description{
Group centering or grand-mean centering the specified variables and optionally add them to the data
}
\examples{
data("news")
res <- centerESM(data = news, subjnr="subjnr",
       addmeans = TRUE, 
       varnames = c("Fearful", "Relaxed", "Anxious"), 
       center = c("person", "person", "person"))
}
