% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNews.R
\docType{data}
\name{DataNews}
\alias{DataNews}
\title{A test dataset}
\format{A data frame with 1727 rows and 83 variables:
\describe{
  \item{subjnr}{subjnr}
  \item{beepnr}{beepnr}
  \item{daynr}{daynr}
  \item{News_YesNO}{News perceived, 0 = NO, 1 = Yes}
  \item{Sudden}{suddenness of news}
  \item{Relevanz}{relevancy of news}
  \item{Dramatic}{dramatic level of news}
  \item{Fearful}{fearful level of news}
  \item{Hopeful}{hopefulness of news}
  \item{Inspiring}{inspiring level of news}
  \item{Cheerful}{feel cheerful}
  \item{Anxious}{feel anxious }
  \item{Down}{feel down }
  \item{Relaxed}{feel relaxed }
  \item{Insecure}{feel insecure }
  \item{Lonely}{feel lonely }
  }}
\usage{
DataNews
}
\description{
description
}
\keyword{datasets}
