% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esmNetwork.R
\name{esmNetwork}
\alias{esmNetwork}
\title{Function to compute network of ESM variables using lags in multilevel analysis}
\usage{
esmNetwork(dat, subjnr, level1, level2 = NULL, vars, covs = NULL,
  randomAll = FALSE, randomVars = NULL, randomIcept = TRUE,
  fixedIcept = TRUE, groups = NULL, lagn = 1,
  centerType = "person", layout = "spring", labs = NULL,
  solid = 0.1, plimit = 0.05, titlePlot = "Figure")
}
\arguments{
\item{dat}{data frame}

\item{subjnr}{variable name, indicating the subjects}

\item{level1}{variable name, indicating level 1 variable, e.g. beeps}

\item{level2}{variable name, indicating level 2 variable, e.g. days}

\item{vars}{vector with the names of the variables for which the network is computed}

\item{covs}{covariates in the analyses, which are not plotted in the network}

\item{randomAll}{logical indicates whether all variables should be used as random effects}

\item{randomVars}{vector of variable names, used as random effects}

\item{randomIcept}{logical indicating whther there is a random intercept (TRUE) or not (FALSE)}

\item{fixedIcept}{logical indicating whether there is a fixed intercept (TRUE) or not (FALSE)}

\item{groups}{variable used to label groups in the network figure}

\item{lagn}{number of lags used in the network}

\item{centerType}{vector of same length as vars that indicates for each variable which centering must be,
applied, either "person" or "grand_mean" or a number. If only "person" or "grand_mean" is specified (default), 
all variables are person or grand-mean centered, respectively.
If centerType is NULL no variables are centered.}

\item{labs}{labels used in the network plot}

\item{solid}{effect size above which lines are shown as solid (default = .10)}

\item{plimit}{p-value under which lines are shown (default = .05)}

\item{titlePlot}{title for the plot}
}
\value{
a qgraph object (network) that can be plotted
}
\description{
Function to compute network of ESM variables using lags in multilevel analysis
}
\examples{
data("DataNews")
vars <- c("Fearful","Hopeful","Anxious","Down","Irritated","Relaxed","Insecure")
labs <- c("FF","HO","ANX","DOW", "IRR","REL","INS")
res <- esmNetwork(dat = DataNews, subjnr="subjnr",level1 = "beepnr",
                  level2= "daynr", vars = vars, labs = labs, lagn = 1)
       
}
