% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratitude.R
\docType{data}
\name{gratitude}
\alias{gratitude}
\title{A test dataset}
\format{A data frame with 7194 rows and 11 variables:
\describe{
  \item{subjnr}{subjnr}
  \item{beepno}{beepnr}
  \item{dayno}{daynr}
  \item{wellBeing}{well being}
  \item{group}{group indicator based on low or high levels of well being}
  \item{pa}{positive affect 3x}
  \item{na}{negative affect 3x}
  }}
\usage{
gratitude
}
\description{
description
}
\keyword{datasets}
