% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagESM.R
\name{lagESM}
\alias{lagESM}
\title{Function to compute lagged n variables in ESM data and add it to data}
\usage{
lagESM(data, subjnr = "subjnr", level2 = NULL, level1 = "beepnr",
  lagn = 1, varnames = NULL)
}
\arguments{
\item{subjnr}{variable, indicating subjects}

\item{level2}{variable, indicating level 2 variable, e.g. days. Can be left empty.}

\item{level1}{variable, indicating level 1 variable, e.g. beeps. Must be specified.}

\item{lagn}{number of lags}

\item{varnames}{vector with the names of the variables for which lagged versions are computed}

\item{dat}{data frame}
}
\value{
data frame with lagged variables added, with suffix L1, L2, L3, respectively
}
\description{
Function to compute lagged n variables in ESM data and add it to data
}
\examples{
data("DataNews")
res <- lagESM(data = DataNews, subjnr="subjnr", level2= "daynr",
       level1 = "beepnr", lagn = 1, varnames = vars)
}
