% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\docType{data}
\name{news}
\alias{news}
\title{A test dataset: news}
\format{
A data frame with 654 rows and 13 variables:
\describe{
  \item{subjnr}{subjnr}
  \item{daynr}{daynr}
  \item{beepnr}{beepnr}
  \item{gender}{gender}
  \item{age}{age}
  \item{Fearful}{fearful level of news}
  \item{Hopeful}{hopefulness of news}
  \item{Inspiring}{inspiring level of news}
  \item{Irritated}{feel irritated}
  \item{Anxious}{feel anxious }
  \item{Down}{feel down }
  \item{Relaxed}{feel relaxed }
  \item{Insecure}{feel insecure }
  }
}
\usage{
news
}
\description{
description: perception of daily news items
}
\keyword{datasets}
