% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDifCor.R
\name{permDifCor}
\alias{permDifCor}
\title{Test the difference of the network connectivity between two groups based on the correlations}
\usage{
permDifCor(dat, vars, group, perms = 1000)
}
\arguments{
\item{dat}{data frame}

\item{vars}{vector with the names of the dependent variables.}

\item{group}{dichotomous variable that indicates the two groups to be compared}

\item{perms}{number of permutations.}
}
\value{
Estimate of difference between groups wrt correlation based network connectivity with p value based on permutations.
         The p-values for differences of all individual correlations and of summaries are given.
         Also p-values for the differences of the centrality measures outDegree and closeness and betweenness are given.
}
\description{
Note that these variables must exist in the data.
}
\examples{
data("gratitude")
vars <- c("pa_1","pa_2","pa_3","na_1","na_2","na_3")
out <- permDifCor(dat=gratitude,vars=vars, group="wellBeing", perms = 10)
}
