% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permfunc.R
\name{permfunc2}
\alias{permfunc2}
\title{Test the difference of the network paths between two groups}
\usage{
permfunc2(iter, dat, pred, dv, group, subjnr, nobs.per.person,
  group.per.person, is.conti = TRUE)
}
\arguments{
\item{iter}{number of permutations}

\item{dat}{data frame}

\item{pred}{vector with names of predictor variables}

\item{dv}{name dependent variable}

\item{group}{dichotomous variable that indicates the two groups to be compared}

\item{subjnr}{identification number of the subjects}

\item{nobs.per.person}{vector with number of observations per person}

\item{group.per.person}{vector with group number for each person}

\item{is.conti}{boolean to indicate if dependent variable is continuous or dichotomous (NOT ACTIVE)}

\item{outnames}{vector with the names of the variables which are centered}

\item{pred}{formula with names of predictor variables for lmer}
}
\value{
matrix (4 x 3) with total differences, diagonal differences, off-diagonal differences,
        and differences between standard deviations, with their p-values (2 definitions).
}
\description{
The function uses permutations to obtain p-values of the path differences.
}
