% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.permDif.R
\name{plot.permDif}
\alias{plot.permDif}
\title{Function to plot the results of the conDif function for network connectivity 
differences between two groups}
\usage{
\method{plot}{permDif}(x, ...)
}
\arguments{
\item{x}{result of permDif}

\item{...}{additional parameters}
}
\value{
Four figures are build with ggplot and returned to the plot window
}
\description{
There are five lot types. The type number indicates what figure is plotted, 
           1 = mean differences total
           2 = mean differences auto-regression, 
           3 = mean differences, excluding auto regression,
           4 = mean differences, subset of predictors
           5 = mean differences of SD's
}
