% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubject.R
\name{plotSubject}
\alias{plotSubject}
\title{Network plot of an individual}
\usage{
plotSubject(res, subjectID, layout = "circular", solid = 0.1)
}
\arguments{
\item{res}{result of class "esmNetwork"}

\item{subjectID}{identification number that indicates the subject to be plotted}

\item{layout}{layout of the plot (see documentation of qgraph)}

\item{solid}{effect size above which lines are shown as solid (default = .10)}
}
\value{
an individual qgraph plot (network)
}
\description{
Network plot of an individual
}
\examples{
\dontrun{
plotSubject(res = res, subjectID = 31623)
}

}
