% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerESM.R
\name{centerESM}
\alias{centerESM}
\title{Function to compute lagged n variables in ESM data and add it to data}
\usage{
centerESM(data, subjnr = NULL, addmeans = TRUE, varnames = NULL,
  center = NULL)
}
\arguments{
\item{subjnr}{variable, indicating subjects}

\item{addmeans}{logical indicating whether a vector of person means should be added to the data 
(works only if person centering is requested for that variable).}

\item{varnames}{vector with the names of the variables which are centered.}

\item{center}{vector of same length as varnames that indicates for each variable which centering must be,
applied, either "person" or "grand_mean" or a number. If "person" or "grand_mean" is specified once (default), 
all variables are person or grand-mean centered, respectively. 
If center is NULL no variables are centered and a warning is issued.}

\item{dat}{data frame}

\item{level2}{variable, indicating level 2 variable, e.g. days. Can be left empty.}

\item{level1}{variable, indicating level 1 variable, e.g. beeps. Must be specified.}
}
\value{
data frame with centered variables, and if requested, person means with suffix "_means".
}
\description{
Function to compute lagged n variables in ESM data and add it to data
}
\examples{
data("DataNews")
res <- centerESM(data = DataNews, subjnr="subjnr",
       addmeans = TRUE, 
       varnames = c("Lonely", "Relaxed", "Anxious"), 
       center = c("person", "person", "person"))
}
