% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conDif.R
\name{conDif}
\alias{conDif}
\title{Test the difference of the network connectivity between two groups}
\usage{
conDif(dat, vars, group, subjnr, level1, level2 = NULL, randomVars = F,
  perms = 500, optim = "bobyqa")
}
\arguments{
\item{dat}{data frame}

\item{vars}{vector with the names of the variables which are centered.}

\item{group}{dichotomous variable that indicates the two groups to be compared}

\item{subjnr}{identification number of the subjects}

\item{level1}{variable, indicating level 1 variable, e.g. beeps. Must be specified.}

\item{level2}{variable, indicating level 2 variable, e.g. days. Can be left empty.}

\item{randomVars}{logical, indicating whether the variables should be included as random effects.}

\item{perms}{number of permutations.}

\item{optim}{optimizer used in lmer, options: "bobyqa" or "Nelder_Mead", see lmerControl (lme4)}
}
\value{
Estimate of difference between groups wrt network connectivity with p value based on permutations.
           and permutation distribution
}
\description{
Test the difference of the network connectivity between two groups
}
\examples{
data("gratitude")
vars <- c("pa_1","pa_2","pa_3","na_1","na_2","na_3")
out <- conDif(dat=gratitude,vars=vars, group="wellBeing", subjnr="subjnr",
level1="beepno", level2 = "dayno", randomVars = FALSE, perms = 10) 
}
