% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataNews.R
\docType{data}
\name{DataNews}
\alias{DataNews}
\title{A test dataset}
\format{A data frame with 1727 rows and 83 variables:
\describe{
  \item{y1}{dependent}
  \item{x1}{predictor}
  \item{m1}{mediator 1}
  \item{m2}{mediator 2}
  \item{m3}{mediator 3}
  \item{c1}{covariate 1}
  \item{c2}{covariate 2}
  \item{mod1}{moderator 1}
  \item{mod2}{moderator 2}
  \item{bimod1}{dichotomous moderator 1}
  \item{bimod2}{dichotomous moderator 2}
  }}
\usage{
DataNews
}
\description{
description
}
\keyword{datasets}
