% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function esmNetwork.R
\name{esmNetwork}
\alias{esmNetwork}
\title{Function to compute network of ESM variables using lags in multilevel analysis}
\usage{
esmNetwork(dat, subjnr, daynr = NULL, beepnr, vars, covs = NULL,
  lagn = 1, labs = NULL, solid = 0.1, plimit = 0.05,
  titlePlot = "Figure")
}
\arguments{
\item{dat}{data frame}

\item{subjnr}{variable, indicating subjects}

\item{daynr}{variable, indicating days}

\item{beepnr}{variable, indicating beeps}

\item{vars}{vector with the names of the variables for which the network is computed}

\item{covs}{covariates in the analyses, which are not plotted in the network}

\item{lagn}{number of lags used in the network}

\item{labs}{labels used in the network plot}

\item{solid}{effect size above which lines are shown as solid (default = .10)}

\item{plimit}{p-value under which lines are shown (default = .05)}

\item{titlePlot}{title for the plot}
}
\value{
a qgraph object (network) that can be plotted
}
\description{
Function to compute network of ESM variables using lags in multilevel analysis
}
\examples{
data("DataNews")
res <- esmNetwork(dat = DataNews, subjnr="subjnr", daynr= "daynr",
       beepnr = "beepnr", vars = vars, labs = labs, lagn = 1)
       
}
