% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function LagESM.R
\name{LagESM}
\alias{LagESM}
\title{Function to compute lagged n variables in ESM data and add it to data}
\usage{
LagESM(dat, subjnr = "subjnr", level2 = NULL, level1, lagn = 1,
  varnames)
}
\arguments{
\item{dat}{data frame}

\item{subjnr}{variable, indicating subjects}

\item{level2}{variable, indicating level 2 variable, e.g. days}

\item{level1}{variable, indicating level 1 variable, e.g. beeps}

\item{lagn}{number of lags}

\item{varnames}{vector with the names of the variables for which lagged versions are computed}
}
\value{
data frame with lagged variables added, with suffix L1, L2, L3, respectively
}
\description{
Function to compute lagged n variables in ESM data and add it to data
}
\examples{
data("newsData")
res <- lagESM(dat = newsData, subjnr="subjnr", level2= "daynr",
       level1 = "beepnr", lagn = 1, varnames = vars)
}
