% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permDif.R
\name{permDif}
\alias{permDif}
\title{Test the difference of the network connectivity between two groups}
\usage{
permDif(dat, vars, covs = NULL, group, subjnr, randomVars = NULL,
  subset = NULL, perms = 500, optim = "bobyqa")
}
\arguments{
\item{dat}{data frame}

\item{vars}{vector with the names of the dependent variables.}

\item{covs}{vector with the names of the covariates.}

\item{group}{dichotomous variable that indicates the two groups to be compared}

\item{subjnr}{identification number of the subjects}

\item{randomVars}{vector, indicating which variables should be included as random effects. 
If "all" then all fixed effects are taken. If "null" only intercept is used as random effect.}

\item{subset}{subset of predictor variables which are compared in summary statistics. If null then result is also null.}

\item{perms}{number of permutations.}

\item{optim}{optimizer used in lmer, options: "bobyqa" or "Nelder_Mead", see lmerControl (lme4)}
}
\value{
Estimate of difference between groups wrt network connectivity with p value based on permutations.
         and permutation distribution. The p-values for differences of all individual paths and of summaries are given.
}
\description{
Test the difference of the network connectivity between two groups
}
\examples{
data("gratitude")
vars <- c("pa_1","pa_2","pa_3","na_1","na_2","na_3")
out <- permDif(dat=gratitude,vars=vars, group="wellBeing", subjnr="subjnr",
randomVars = FALSE, perms = 10) 
}
