% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permfunc.R
\name{permfunc}
\alias{permfunc}
\title{Function to test the difference of the network connectivity between two groups}
\usage{
permfunc(perms, dat, pb, outnames, pred, nobs.per.person, group.per.person)
}
\arguments{
\item{perms}{number of permutations}

\item{dat}{data frame}

\item{pb}{name progressbar}

\item{outnames}{vector with the names of the variables which are centered.}

\item{pred}{formula with names of predictor variables for lmer.}

\item{nobs.per.person}{vector with number of observations per person}

\item{group.per.person}{vector with group number for each person}
}
\value{
matrix (4 x 3) with total differences, diagonal differences, off-diagonal differences,
        and differences between standard deviations, with their p-values (2 definitions).
}
\description{
Function to test the difference of the network connectivity between two groups
}
