% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.conDif.R
\name{plot.conDif}
\alias{plot.conDif}
\title{Function to plot the results of the testCF function for network connectivity 
differences between two groups}
\usage{
\method{plot}{conDif}(x, type = 1)
}
\arguments{
\item{x}{result of conDif}

\item{type}{number that indicates what output to plot, 
1 = mean differences total
2 = mean differences auto-regression, 
3 = mean differences, excluding auto regression,
4 = mean differences of SD's}
}
\description{
Function to plot the results of the testCF function for network connectivity 
differences between two groups
}
