% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigConnect.R
\name{sigConnect}
\alias{sigConnect}
\title{Test the difference of the network path strenghts between two groups}
\usage{
sigConnect(dat, vars, group, subjnr, level, randomVars = F,
  perms = 500, optim = "bobyqa")
}
\arguments{
\item{dat}{data frame}

\item{vars}{variables in network}

\item{group}{dichotomous variable that indicates the two groups to be compared}

\item{subjnr}{identification number of the subjects}

\item{level}{variable indicating level (e.g. days or beeps)}

\item{randomVars}{boolean indicating whether the variables are random effects in the model}

\item{perms}{number of permutations}

\item{optim}{optimizer used in lmer, options: "bobyqa" or "Nelder_Mead", see lmerControl (lme4)}
}
\value{
matrix (k x k) with observed differences of the paths ( k = number of variables in network), 
        2 matrices (k x k) with the p-values (2 definitions) obtained from permutation
        distribution for the differences. A for the number of tests adjusted alpha is given for the nominal 5%.
}
\description{
Function that uses resampling method (permutations) to obtain distribion and p-values for
the differences of the path strengths of a network between two groups.
}
\examples{
data("gratitude")
vars <- c("pa_1","pa_2","pa_3","na_1","na_2","na_3")
sigConnect(dat = gratitude, vars = vars , group="wellBeing", subjnr ="subjnr", 
level = "beepno", randomVars = FALSE, perms = 10)
}
